renamed `ProjectList` component to `ProjectsNav`
