+++ src/routes/wheres-waldo.lazy.tsx
@@ -1,6 +1,6 @@
 import { createLazyFileRoute } from '@tanstack/react-router'
 import Gallery from '@/components/Gallery'
-import ProjectList from '@/components/ProjectList'
+import ProjectsNav from '@/components/ProjectsNav'
 import {
     typeScript,
     react,
@@ -96,7 +96,7 @@ function wheresWaldo() {
                 />
             </article>
 
-            <ProjectList filter="wheresWaldo" className="bg-red-100" />
+            <ProjectsNav filter="wheresWaldo" className="bg-red-100" />
         </main>
     )
 }
